Added a debugging helper function and a function to enable/disable the output of debug messages. Replaced all calls to printf() with calls to axgsq_error(). A few other minor fixes as well.
It will be interesting to see what happens when the extension is used on in a web server.  As the axgsq_debug() function for enabling/disabling debug output uses a global var to store the status.
Does php close/re-open on each file parse or stay open and carry them over?
On a multi-user/site server does the debug option affect every one?
Can you enable debug on one page and it affect another page?